implemented data cleanup

+++ copper/machines/rhythms_multiplied.py
@@ -21,7 +21,7 @@ class RhythmsMultiplied(object):
     def set_logical_tie(self, logical_tie, **kwargs):
         super().set_logical_tie(logical_tie, **kwargs)
         segment_index = logical_tie.parent.parent.my_index()
-        logical_tie.ticks = logical_tie.ticks * self.rhythm_multipliers[segment_index]
+        logical_tie.ticks = int(logical_tie.ticks * self.rhythm_multipliers[segment_index])
 
     def set_segment(self, segment, **kwargs):
         super().set_segment(segment, **kwargs)
